fix: notification new season
