+++ sseudam-clients/notification/src/main/kotlin/com/sseudam/client/discord/DiscordMessagePayload.kt
@@ -1,0 +1,5 @@
+package com.sseudam.client.discord
+
+data class DiscordMessagePayload(
+    val content: String,
+)
