+++ sseudam-clients/notification/src/main/kotlin/com/sseudam/client/discord/DiscordWebhookClient.kt
@@ -13,7 +13,3 @@ interface DiscordWebhookClient {
         @RequestBody message: DiscordMessagePayload,
     )
 }
-
-data class DiscordMessagePayload(
-    val content: String,
-)
