+++ sseudam-core/core-api/src/main/kotlin/com/sseudam/presentation/v1/visit/VisitedController.kt
@@ -28,8 +28,7 @@ class VisitedController(
         user: User,
         @PathVariable spotId: Long,
     ): SpotVisitedDetailResponse {
-        val spot = trashSpotService.findBy(spotId)
-        val result = spotVisitedFacade.visitSpot(user.id, spotId, spot.suggesterId)
+        val result = spotVisitedFacade.visitSpot(user.id, spotId)
         return SpotVisitedDetailResponse.of(
             isToday = result.isToday,
             visited = result.visited,
